patterns/strands-single-agent/requirements.txt
# Strands agent dependencies with pinned versions
strands-agents==1.24.0
mcp==1.26.0
bedrock-agentcore[strands-agents]==1.2.0
PyJWT[crypto]>=2.10.1
requests>=2.32.0
